joystick: support deferred init and configurable retry interval
The initialization can now allow for a missing joystick. In more recent
kernel versions, the Xbox controller joystick device file is only
created after the joystick is connected, not already if the receiver is
plugged in.

The configurable retry interval allows to determine the time after which
another attempt is made to open the joystick device. Note that this
might just be the delay from a button press to a reaction on the robot.