libconfig: notify change handlers whenever a value is changed in YAML
So far, the change handlers were only notified if a config value was
changed in a YAML file. The YAML config can also be updated by other
means, e.g., by a network client. Also notify the change handlers if the
config was changed by calling one of the set functions.