tf: read data of just opened interfaces on init once
During initialization, we open already existing interfaces. We must
trigger a data changed event for those in order to make sure that we get
their data. Otherwise, already existing static transform interfaces
would not properly update, as they are written only once. Most likely
during their initialization.

While this is not a problem for plugins -- as the TransformListener is
initialized well before anything would publish transforms -- it can be
if running the TransformListener externally.