tf listener: do not use writer information if blackboard is remote
If the blackboard is a remote blackboard, we cannot access any writer
information, because this is not implemented. For now, add a constructor
argument bb_is_remote, which defaults to false. If it is set to true, no
interface will be closed when the writer disappears, and the writer
authority is set to "remote".