mongodb: better separate client config
Split client configuration class into separate files, rename configs_ to
client_configs_, and move initialization to separate method.

These are preparations to introduce new kinds of configurations.

NOTE: this also introduces small configuration changes. The "active"
flag is renamed to "enabled". Furthermore, the hosts for a replica set
are now stored in a list rather than a map making use of additions to
the config API.