laser calibration: add roll calibration functionality
The tool is now able to configure the roll rotation of the back laser.
It assumes that there are no obstacles in the back of the laser, and it
adapts the roll rotation of the back laser static transforms such that
the laser points on the ground have the same mean height on the left and
on the right.