laser calibration: add pitch calibration
The calibration expects the z transform of the laser to be correct. It
changes the pitch of the laser until the mean height (z value) of all
points in the rear of the robot is 0.

Note that pitch and roll calibration depend on each other. To get better
results, run the calibration multiple times.