laser calibration: add yaw calibration
The yaw calibration compares front and back laser and matches laser
points of both lasers on the right of the robot such that the total
distance between the points is minimal. In each step, a matching between
the two lasers is computed, which minimizes the current total distance.
The yaw angle is then adapted to reduce the min total cost. All previous
yaw/cost pairs are saved. In each step, if any previous yaw angle had a
smaller cost than the current cost, we randomly decide whether to use
the previous angle or the current angle in the next iteration, where the
probability of using the previous angle is higher if the previous cost
was lower. This way, we can escape local minima while still converging
to a result eventually.