syncpoint: throw an exception if an emitter tries to release a syncpoint
An emitter always needs to call unregister_emitter before releasing a
syncpoint. We could also do this automatically, but we would need to
determine if the component is an emitter for the original syncpoint when
releasing all predecessor syncpoints. As this is error-prone, we expect
the syncpoint user to call unregister first instead.