dynamixel: setup syncpoint usage
The dynamixel plugin consists of an act and a sensor thread. The act
thread is dependant on data generated by the sensor thread. In rare
cases the sensor thread takes more time resulting in situations where
exec_act is called although exec_sensor did not finish yet.
With this commit we make use of SyncPoints to let the act_thread wait
until the sensor thread finished (syncpoint "/sensor/aquire").
After the act_thread is done we emit "/act/exec/end".