mongodb: allow initialization without client connection
This reverts the allow for fail flag, which was a pretty bad idea to
begin with.

Now, there are two ways to initialize the MongoDBAspect:
1. Give a config name
   This works exactly as before, a connection is established and it is
   an error if this fails.
2. Omit config name
   This will not initialize the connection. Only the connection manager
   is initialized and may be used to establish a connection later.

The new API makes more sense, but it is not fully backward compatible.
If the configuration name was omitted previously, it resulted in a
default connection to be established. This has now to be done
explicitly.

This modification is useful especially considerig more complex
connection setups such as replica sets, especially if they are managed
by the mongodb plugin. An RS might simply not be ready during
initialization of the connection, e.g., leader election may take some
time. Therefore, allow to postpone this and leave it to the plugin to
handle this case gracefully.

The previous attempt allowing for initial connection failure was flawed.
For example, if the initial connection failed, there was no way to use
the configured authentication data.