clips-executive: CLIPS modularized version
This version uses CLIPS modules to separate the code. However, initial
testing suggests that this does not bring the expected benefit, to the
contrary, it makes things harder, more fragile, and requires more code
for explicit flow handling.

This version is for reference only. We will continue without modules for
the time being.