clips-executive: add domain representation
Add a representation for a planning domain with operators, actions,
and preconditions. An action is an instantiated operator, i.e., an
operator with a parameter grounding, which assigns all parameters of
the operator to objects of the domain.

This currently allows to represent STRIPS domains with typing. Effects
are not yet represented.