clips-executive: add rules for grounding and precondition checking
These rules ground all atomic preconditions according to a grounding in
the fact base and compute which preconditions are satisfied. Grounded
preconditions are added as new facts to the fact base, so we always
have the ungrounded, parameterized version of a precondition in
addition to the grounded precondition. Only grounded preconditions are
tested for satisfaction. Preconditions are checked one by one, starting
with the atomic preconditions, and continuing with negations and
conjunctions.