clips-executive: split parameters into params and param-values
Instead of having a list of key-value pairs as a single multi-field
slot, separate parameter names and values into two separate slots.
This allows an easier handling of parameters when checking the
precondition of the action.

For now, we just reconstruct the old list when passing the arguments to
the skiller.