clips-executive: check precondition using params from the plan-action
Instead of having separate grounding facts, use the grounding that is
defined by the plan-action. Also compute the grounding of all
parameters of each (atomic) precondition in one run instead of
substituting them one by one. This also changes the parameter
representation to the same representation as in the plan-action
(separate parameters and parameter values).