clips-executive: add support for predicates with constants
Sometimes a precondition contains a constant rather than a parameter of
the operator. For this case, add a multislot param-constants to the
domain-atomic-precondition template. The multislot defines the constant
values to ground the predicate with. If the ith value of
param-constants is some value v != nil, the ith value of param-values
will be v after grounding. If the ith slot is nil or does not exist,
the precondition will be grounded with the grounding of the ith
param-name, as before. Note that in the case that a constant is set for
a slot, the respective slot of param-names will be ignored.