clips-executive: adapt test scenario to use a proper domain
The domain consists of the two operators say-hello and say-goodbye.
Both operators do not have any parameters. The operator say-hello is
always executable (its precondition is true), the operator say-goodbye
can only be executed if hello has been said before. The operators
say-hello and say-goodbye have the effects that the texts hello and
goodbye are said respectively.