clips-executive: distinguish domain-predicate and domain-fact
Rename the existing domain-predicate template to domain-fact. It now
describes instantiations of a predicate. The predicate itself is now
described by the domain-predicate template with a name and a set of
possible parameters. Along the way rename the parameters slot to the
more precise param-values to be in line with other definitions.

Adapt the domain, example, and test code where necessary to account
for the change.