clips executive: add interface to PDDL generator and planner
This allows to generate a PDDL problem and then call a PDDL planner to
compute a plan. Calling the function pddl-call with a specified goal
will trigger the process. The current status of the planner can be
checked with the pddl-plan deftemplate.

Note that this does not yet pass the given goal to the planner, nor does
it translate the resulting plan back into CLIPS.