clips-executive: add support for interface template preloading
For each blackboard interface type to use a template is created. This
usually happens when opening an interface. This can be cumbersome and
require to split code for opening and processing interfaces (as the
template for processing is not yet available at load time).

To avoid this splitting add a rule that can preload interface type
templates based on a spec-specific config value. The functionality in
the CLIPS blackboard feature was already there, just not exposed in the
executive.