clips-executive: also initialize robot memory when using PDDL planner
When we are using a PDDL planner, the resulting plan is stored in robot
memory. Therefore, initialize robot memory during init of the executive
and register a trigger for new plans.

On the way, rename the feature to pddl_planner, as the CLIPS feature
for the PDDL parser is already called pddl.