clips-executive: fetch the PDDL plan from robot memory and expand goal
After calling the PDDL problem generator and the planner and then
waiting for the planner to finish, fetch the resulting plan from robot
memory and assert plan actions for all actions in the plan. This
expects the goal to still be SELECTED and sets the goal to EXPANDED.

To make sure the plan is indeed for the selected goal, the message id
of the plan message is tracked and associated with the goal. Only a
matching plan in the database is used to expand the goal. This allows
us to expand multiple goals simultaneously in the future.

Note that this does not set the param-names of the plan-actions,
because the names of the parameters are not encoded in the plan. Thus,
this needs to be taken care later by checking the PDDL operators for
the correct parameter names.