clips-executive: switch to PDDL-based domain for the test scenario
This is more or less the same scenario as before, but now uses a PDDL
planner to create the plan instead of specifying a fixed sequence.

In this scenario, the goal is to say goodbye to peggy. In order to do
that, the agent first needs to say hello to peggy. The resulting plan
should have two actions: say-hello(peggy), say-goodbye(peggy, goodbye).