clips-executive: prioritize wm meta data sync and check
Set ID or key with highest priority, also check for conflicts with
higher priority to catch them as soon as they happen (if they happen).
Must come after setting key/ID or otherwise (if multiple facts are added
in a single rule) may fire matching "" or () for ID and key
respectively.