clips-executive: assume skill exec only if mapping available
Make action execution a more generic concept. Allow actions other than
skills to be executed. A first step is to pre-load known action skill
mappings and only trigger execution for actions which have such a
mapping. Accordingly, rename plan-exec.clp to skills-actions.clp as this
is now considered one specific set of actions, and not the only plan
execution pipeline.