clips-executive: do not apply effects for sensed predicates
Sensed predicates are predicates that are updated by the system using
its sensors, e.g., a camera seeing an object, or a gripper with a puck
detection. For these predicates, do not update the domain predicate for
actions that modify the predicate. Instead, we expect the sensors to do
this.