syncpoint: remove waiter on syncpoint release
If a component is currently waiting when it is releasing the syncpoint,
we need to remove the component from the waiting list so it can later
get the syncpoint and wait again.

Also reset the timer if the releasing component was the component that
started the timer. Note that may cause other components to wait
indefinitely, even in a timed wait, because the timer was stopped.