clips-executive: add maintenance goals
Goals can now have two types: ACHIEVE and MAINTAIN. ACHIEVE is the
a usual goal that needs to be achieved once to be completed. MAINTAIN is
a goal that does not have actions itself, but instead ACHIEVE goals as
sub-goals. These ACHIEVE sub-goals need to be created as part of the
goal lifecycle of the MAINTAIN goal.