pddl-robot-memory: check for duplicate names in problem template
For each query in the problem template, we create a dictionary entry
with the template name as key and the result of the query as value.
If the same name is used twice in the template, then we create duplicate
entries in the resulting problem. To avoid this, remember all names and
their respective queries. If a name occurs again with the same query,
skip querying the database and use the result from the previous query.
If the queries are different, print an error message.