clips-executive: domain to world model fact ID remapping support
This allows to setup re-mapping for domain facts to arbitrary world
model fact IDs. This is necessary to achieve a better separation of the
domain and world models by avoiding domain-model specific naming for
world model facts.

Currently, the predicate name can be re-mapped to an arbitrary wm-fact
ID path (i.e., the prefix in front of args?). Other possibilities for
the future could be argument re-mapping and argument value translation.

Mapping can be enabled by either asserting appropriate
wm-sync-remap facts, or using the wm-sync-remap-id-prefix utility
function for simpler mappings (path-prefix re-mapping only, the last
path element is the same as the predicate name in the domain.