clips-executive: atomic processing of effects
The effects were previously handled such that a rule fired once for each
effect. However, this was prone to a race condition. Consider an action
with effects on a sensed and a non-sensed predicate. Further, assume
that the sensed predicate effect was applied first. This would assert an
domain-pending-sensed-fact fact. If, however, the expected value was
sensed in the meantime, the fact would be removed immediately. This
would then trigger domain-action-is-final, which only checks for the
absence of pending sensed facts. On our example, the non-sensed
predicate effect would not be applied.

We avoid this issue by reducing the effect application to a single
domain-apply-effects rule which asserts all actual and pending effects
at once. The only alternative would be a very elaborate final check that
makes sure that all effects have been applied.

The new rule is based directly on the previous method, also including
the conditional effects, which would deserve some documentation.

This issue was reported by Daniel Habering with a useful demonstration
case.