clips-executive: support ignoring sensed effects for an operator
Sometimes, there is a slight discrepancy between the model as required
during planning and execution. In that case, it is necessary to ignore
sensed effects for a specific domain operator.

For example, in the RCLL, exogenous machine actions are modeled within
the domain to assert the appropriate effects during planning. For
example, for the cap station, an action models the retrieval of the cap.
It expects the machine to be in processing state, which is an effect of
the action to put a workpiece on the machine. The predicate representing
the state is a sensed predicate. Now, the processing state can appear
only shortly, such that waiting for it in the putting action can result in
a deadlock as the expected intermediate state is not observed. Having
this intermediate state is necessary for planning. However, in execution
just ignoring the sensed effect in this situtation solves the deadlock.
Therefore, for effects which may only very shortly be observed during
execution this is a useful option.

We might later want to make this more specific, i.e.,, allow to declare
for an operator for which sensed predicates not to wait. Or even
further, to generally declare which predicates to ignore and not assert.
This would be similar to the effect handling the the ROSPlan pipeline of
the RCLL.