clips-executive: fix action IDs for PDDL actions
The last method did not work because the plan-action fact was not
asserted yet before adding the next action, which resulted in same
action IDs for all actions in a plan. Now, get the highest existing
action ID only once as offset for the action IDs. Set the final action
ID to the sum of the offset and the action ID in the plan.