clips-executive: avoid using domain-retracted-fact in RM upsync
The code to update the robot memory representation of the domain model
used (domain-retract-fact)s to determine when to remove a fact. However,
domain-retracted-fact and delayed application pose their own class of
issues, especially during effect application (see following commit).

Therefore, avoid using domain-retracted-fact by introducing an new
robot-memory-sync-mapped fact for each domain-fact we have inserted into
the robot memory. This can then be used to determine when we have to
remove information from robot memory.

This has the negative side effect to increased (double) the required
memory and clutter the working memory. This is unfortunate, but at the
moment the lesser evil. Once we implement bidirectional sync we probably
need something like this anyway.