clips-executive: remove domain-retracted-fact in domain model
Remove the deftemplate and the now superfluous retract and cleanup
rules. Additionally, update the effect application to directly retract
(domain-fact)s.

The original problem is in sequencing the application of effects, i.e.,
to apply them in the order they are stated in the action operator. In
PDDL, for example, it is possible to state conflicting effects and still
get the desired output. For example, for a (move A A), the effect could
be (and (not (at a)) (at a)). While logically, this contradiction is
invalid, planners simply apply them in sequence yielding a valid fact
base of (at a) afterwards.

A first try was to sequence the effect applications. However, the
delayed retraction can still cause problems, i.e., if the rule to apply
the retraction or cleanup of a retracted fact f1 fires before the
sequence of applied effects already included this very fact f1.

The solution is to modify the overall system to no longer rely on
(domain-retracted-fact) and therefore be able to immediately retract it,
thereby keeping the sequence mandated by the effect specification and
replicating the PDDL behavior.

This was discussed with Mostafa Gomaa to come up with the solution.