clips-executive: action IDs are only unique for a fixed goal and plan
Action IDs are no longer unique globally, but  for a given goal and
plan. Similarly, plan IDs are no longer unique globally, but only for a
given goal.

Adapt precondition grounding and checking and effect application such
that it always checks the goal ID and plan ID. All plan-actions now
always need to refer to a goal and a plan.

Also modify the test scenario and all tests to the change and add a test
that checks that we only apply effects for the right action, even if
there is a second action with the same action ID.