clips-executive: add the operator to domain preconditions
In some cases, we need to refer to the operator of a nested/atomic
precondition. Add a slot operator that always gives the name of the
operator the precondition belongs to. For the root precondition (i.e.,
the top-most precondition), the value of the slot operator is the same
as the value of the slot part-of.