clips-executive: use salience for prioritized wm sync
A problem an arise on a dependency chain like this:
- domain model: update a domain-fact
- wm sync: update a wm-fact accordingly
- goal reasoner: base decision on wm-fact

Here, the goal reasoner could act on outdated information as the rule
may fire before the wm-sync is complete. Fix this by giving wm-sync
operations a higher salience. This ensures that all syncing is complete
before acting on it.

Thanks to Mostafa Gomaa for reporting this issue.