clips-executive: fix test for nonunique action IDs
The test expected unstack not to be executable, when in fact it was. The
test is supposed to check that preconditions are properly matched to
actions, even if actions have the same IDs but different goals/plans.
Do the test with stack instead, as stack is in fact not executable in
the given situation.