clips-executive: add exogenous actions, drop precond on sensed effect
For any exogenous action, a precondition is removed from the operator
definition if the operator also has the same predicate with the same
parameters and constants as sensed effect. This means that parts of the
effect may already have occurred before the exogenous action is
executed. By modifying the precondition, we make sure that the exogenous
action is still executable in that situation.