pddl-planner: add interface field for planner success
The planner does not always find a plan, so the interface field "final"
does not suffice to determine whether the planner found a plan. Add a
separate field "success" that is true iff the planner has found a plan.
If the planner failed, "final" will be true, but "success" will be
false.