clips-executive: add value predicates that always have a single value
Value predicates are predicates that are used to store a single value
for a set of parameters. They can be used to serve as a replacement for
functions. For a value predicate, all but the last parameter are
considered to be function parameters, while the last parameter of the
predicate is considered to be the value of the function. For given
function parameters, only one value may exist, i.e., the predicate must
be false for all but one value.