clips-executive: fix bogus unknown parameter error
In some cases (e.g., if the plan is generated with a PDDL planner), the
param-names of an action are not set. In this case, we always take the
parameters from the operator and copy it into the action. However, this
can lead to a race condition, where the precondition of the action is
grounded before the action parameter names are set, leading to bogus
unknown parameter errors. To avoid this, make sure that the length of
the parameter names is the same length as the parameter values before
grounding the precondition.