clips-executive: fix missing grounding of conds for conditional effects
Commit 78be2e85 restricted the grounding of preconditions so it would
only ground if the action is FORMULATED, PENDING, or WAITING. However,
for conditional effects, the action is EXECUTION-SUCCEEDED, when the
condition should be checked.

Remove the restriction for atomic preconditions, as these can be
grounded only if their parent is grounded anyway. This properly grounds
the precondition of condinitional effects without resulting in undesired
late grounding for action preconditions.