clips-executive: replace precond by disjunction for value predicates
For non-value predicates, completely drop a (sub) precondition of an
exogenous action if the action has the precondition or its negation as
effect.

For value predicates, if a predicate is a value predicate (i.e., it can
only be true for one value), and an exogenous action has the value
predicate both as precondition and as effect (with a different value),
then replace the precondition by a disjunction of the original
precondition and the effect. Thus, if the effect already took place,
then action is still executable. As an example, consider the following
action with precondition and effects on the value predicate mps-state:

precond: mps-state(mps PROCESSING)
effect:  mps-state(mps READY-AT-OUTPUT)

Then the precondition os modified to the following precondition:
precond: mps-state(mps PROCESSING) OR mps-state(mps READY-AT-OUTPUT).