clips-executive: safer grounding of effects with constants
When grounding an effect, we check for each parameter of the predicate
whether it is a constant of the effect or a parameter of the action.
Instead of checking whether the name of the parameter is 'c', check if
the value of the respective slot in the constants is != nil to decide
whether a particular value is a constant.