restapi-gen: add generator for OpenAPI 3 specs w/ C++17
Usually, for API generation a tool such as Swagger can be used. However,
the available outputs are unsuitable (cpprest) or broken and incomplete
(restbed). Additionally, extending swagger with new templates is a
rather tedious process, and it was easily available only as Docker
image, but not in the most recent version capable of OpenAPI 3.0.0.

Enter restapi-gen. This one-night hack simply reads the OpenAPI 3 spec
(YAML) and uses the Jinja2 template engine to emit the desired files.
The power with this approach is clearly in the templates (and these are
admittedly somewhat involved and not easy to read on first sight). But
they are nevertheless so much easier to adapt than when using Swagger.

This version comes with templates suitable to generate data models
fitted to C++17 (uses std::optional) using RapidJSON for rendering and
parsing JSON documents. It supports API models with polymorpism and all
basic and array data types.

Additional templates may be added later if the need arises, e.g., for
older compilers or other languages. I have some for older C++ versions
without polymorphism (but that already breaks our first example).

Upstream is https://github.com/timn/restapi-gen