buildsys: add tool support for generating REST APIs
REST APIs can be easily build with this tool support with just the
OpenAPI 3 YAML spec available. If the file is named api.yaml, include
rest-api.mk and call "make restapi-models". Note that models are *not*
re-generated on a normal build by default. We expect these APIs to
change rarely. This can reduce build-time dependencies (the generated
code is checked in and can be built, the Python parts are not necessary
then, only for re-generating the models).