doc: add Apache 2 license for TypeScript files
We are building on some Kubernetes Dashboard code which is Apache 2.0.
License looks good, let's use that, too, for TypeScript stuff. Headers
are checked for manual style (apps) and API License style in generated
models.

Apache 2.0 is GPL-compatible and hence we can bundle that code in our
project. Note that when using the code integrated it is actually GPL
code. Only if just Apache 2.0 licensed code is used as a basis for
another project the license is actually Apache 2.0.