clips-executive: more differentiation on world model saliences
wm-fact id/key filling must be prioritized above synchronization.

Furthermore, apply deletes before adds (as a "modify" step). This is ok
because either
1. a rule performs multiple updates on a domain-fact, e.g., a modify.
Then, if the last operation is an assert, we need to keep this update.
If the last is a removal, we need to remove the fact. In either case,
only the last operation must be persisted. However, even this single
eventual state leads to two wm-sync calls, a removal of the old domain
fact, and another one to add the new (updated) domain fact. We must
prioritize the deletion so that we guarantee that the add persists.
This is transitive when performing more than just a single modify to a
particular domain fact. There is one initial state (when the rule fires)
and one final state (after the rule completes). The wm-sync only sees
these two states and hence has to perform two operations.
2. or, two rules perform updates on a domain-fact. This is actually of no
concern to us, since the saliences require that the wm is synced after
the first rule, but before the second rules fires.

Thanks to Mostafa Gomaa for reporting this issue.