webview-frontend: new Angular-based frontend for Webview
This is destined to eventually replace the current webview web interface
entirely. The idea is rather different:
- no more generation of HTML pages in C++ code
- expose all information via a REST APIs
- frontend is developed separately, making use of the REST APIs
- REST APIs are specified using the OpenAPI 3 spec (it's a specifically
  formatted YAML file) to allow for automatic code generation.
- Angular is a full framework that takes away the chores of development
  and allows to focus on getting actual functionality integrated.

For now, the only part that is done is to monitor the CLIPS Executive.
It shows the domain model, goals, and plans, and can auto-update for
monitoring. There are still a few ideas hanging but for now it will help
develop the CLIPS Executive and models. This first step focused on
getting new and required functionality in there first. The ring is open
for more wranglers porting and adding more functionality.

It does take a little effort to get into it, but the concerns (backend,
frontend) are now clearly separated. To start with the frontend, you
need to build from source for now. Once we reach a stable version, we
can consider adding a production build. I'll setup a wiki page
describing the process to run and develop.

Overall this took quite a few nights but it kept me sane while writing
:-)