libwebview: regex-based parsing for multisegment path params
The major change is to support multisegment path parameters (cf.
https://github.com/OAI/OpenAPI-Specification/issues/1459). This is
required, for example, when accessing data where the ID itself is a
path-like identifier.

After reviewing a few opportunities go with the referenced proposal. It
is really simple and adds almost no weight to the spec.

To be able to properly parse the paths now use regular expressions.
First, we parse the path strings and generate a full match regex. This
is then stored and used to identify incoming requests.