webview-frontend: blackboard data access implemented
Data can now be read through the new interface. It allows to open (any
number of) cards, one for each interface. These are shown in a dynamic
grid.

What remains to be done is remembering the last opened interfaces to not
have to re-open them every time when visiting the tab. At least for the
duration for a session. Auto refresh still needs some work, too.