libwebview: enable thread pool support and refactor server
The WebServer class has been restructured to group settings in setup_*
methods. Allow for chaining of calls. Rename SSL to TLS (which is what
it actually is today). Add an extra start() method that avoids having to
provide several specific constructors for the different feature setups.
There is now a single constructor with the minimum required set of
parameters. MDH_start_daemon is now called only in the start() method,
not in any ctor.

A new feature introduced is to support libmicrohttpd thread pools. This
allows to handle multiple requests concurrently and increases
performance. Especially when calling a few REST APIs in parallel this
can improve performance on the frontend considerably.