libwebview: functional handlers by URL patterns
This fundamentally (and non-backwards-compatible) changes the way
components interact with webview.

Before: URL prefixes are registered with the URL manager providing a
WebRequestProcessor instance. This has exactly one method that is called
for any incoming request with the given prefix. Path argument parsing
must be handled in the request processor.

Now: The URL manager gets a functional interface allowing to register
handler functions for URL patterns. URL patterns may contain segment
matchers such as {var} (shortest match of anything that does not contain
a slash, typical use is "/path/{id}" to identify an object), {var+}
(shortest match that may also contain a slash, effectively allows to
have IDs which contain a slash itself or to gather larger areas), or
{var*} (greedy match that contains any character). Instead of a single
processor for a prefix, one can now give detailed URL specifications and
register a handler function for each. This eliminates URL parsing in
components for the most part.

The URL manager now holds a router which provides the matching
functionality, similar to the WebviewRestApi.