libbaseapp: thread manager call finalize before aspect finalization
When removing a single thread via the ThreadCollector in the
ThreadProducerAspect, that very thread's finalize() was called after the
aspects had already been finalized. This can (and most likely will) lead
to segfaults, e.g., when trying to close a blackboard interface. fvbase
was indeed suffering from this problem.

This is similar to the fix for thread lists (e71b76e3 from 2013).