libwebview: handle cross-origin resource sharing (CORS)
Cross-Origin Resource Sharing (CORS) is a mechanism that uses additional
HTTP headers to let a user agent gain permission to access selected
resources from a server on a different origin (domain) than the site
currently in use.
(cf. https://developer.mozilla.org/en-US/docs/Web/HTTP/CORS)

We need to handle this properly for two scenarios in particular:
- running a development build of the new frontend not served through
  webview server
- accessing multiple robots from a single app instance (later)

This requires:
- configuration options to setup CORS to a) allow all access from
  anywhere, b) allow specific origins (e.g., all of our robots),
  or c) make no exception and require same-origin access.
- handling OPTION preflight requests

This is now fully implemented. For now, OPTION requests are always
processed by the dispatcher and not passed on to handlers.