webview: add support for explicit high-weight 404
Support the configuration of URLs for which a 404-handler with a very
high weight should be registered. This will give a 404 for non-existent
plugins rather than the frontend. This is useful for endpoints which are
not user-facing, but rather are ingested by some other program. A 404 in
that case can be more helpful than the index file.