webview-frontend: configure default prometheus url only in dev env
For the production deployment do not set a prometheus URL but rely on
the backend info api to provide the URL. It is likely that in production
the prometheus URL is proxied, e.g., to add TLS.

The, however, accessing a wrong URL may lead to issues installing the
service worker. Therefore, do not give a URL at all.

Add a method to query whether a URL is known for a given service.