webview-frontend: smaller style changes
This vertical sizing is driving me crazy. Anyone into CSS styling who
knows how to scale an SVG to the size of the enclosing container,
keeping the scale to 1, overflow with scroll bars, and doing this
without some fixed pixel size in the hierarchy? I most of the time end
up with containers size 0, oversized, or max width correct but then
vertically too large...