webview-frontend: preserve lockout state
The lockout state is now preserved even after closing a tab or
restarting the browser. This is to avoid in hectic in-game situations to
forget to re-enable the lockout.

Maybe the lockout is enough to convince referees to be allowed to
investigate the system in a little more detail while running.