buildsys: use conservative build options by default
-mtune=native can lead to memory alignment issues with the Eigen library
when data is shared with other code that was built with different
optimization. The PCL lib is one such case.

Also replace -O3 with -O2 since some of those optimizations have been
found to be buggy (e.g. uninitialized quaternions in the tf tolua
interface).

Potentially dangerous optimization should not be set by the
domain-independent default repo, but rather by the domain repo. The
infrastructure to do this is already in place, anyway.