laser-filter: add copy filter
The copy filter does not modify the data and can be used to copy laser
data from one interface to another. This can be used to provide
different filter configurations that always result in the same topic
names independent of the hardware. In particular, this allows to specify
a filter that copies data from '$laser' to '$laser 360' to allow simple
switching between 360° lasers and 1080° lasers.