clips-executive: always check for goal before grounding or applying effects
In the context of specifying action ids uniquely ONLY within a given
goal and plan.

Add the missing checks for the existence of the goal specified in the
plan-action during grounding and other similar rules that process
plan-actions or preconditions.

The missing check caused an un-unified behavior, in case the goal-id
was missing in the plan-action. Some of the rules where still allowed to
fire while others needed the goal-id.

(Example, preconditions were still allowed to ground, effects
were allowed to apply. Whereas, atomic/nested precondition were not allowed to
ground)