mongodb: remove unconfigured replica set members
After a configuration change, the replica set may still contain members
that are not configured to be part of the replica set anymore. If that
host was still reachable, it was kept in the replica set. We should only
keep hosts in the replica set that are set as members in the
configuration.

This also fixes a problem when changing the host in the configuration
from the actual hostname to 'localhost'. As they both refer to the same
host, the actual hostname will be kept in the replica set, which will
later cause an error, because we expect either all hosts to be local, or
none of them.