libconfig: add support for sequences of maps to YAML config
The YamlConfiguration supported only maps, scalars, and lists, which are
sequences of strings.

With this change, we allow to have sequences of maps, which allows
something like this:
values:
  - name: x
    some_val: y
  - name: second
  - name: foo bar

These will end up in the following leafs:
/values/0/name: x
/values/0/some_val: y
/values/1/name: second
/values/2/name: foo bar

This slightly richer way can make configuration representation clearer,
more concise, and with an ensured ordering (compared to the previous
workaround to have a map with names for entries rather than numerical
IDs).